- mtcpinterface.cpp :  Keep mtcp_process_pthread_join only if PTRACE defined    (Eventually, remove it for PTRACE too, when satisfied it doesn't hurt.) - various minor cleanups based on errors seen at NMI for various architectures

+++ dmtcp/jalib/jsocket.cpp
@@ -137,7 +137,7 @@ jalib::JSockAddr::JSockAddr ( const char* hostname /* == NULL*/,
     }
 
   } else { // else (hostname, port) not valid; poison the port number
-    _addr[0].sin_port = -2;
+    _addr[0].sin_port = (unsigned short)-2;
   }
 
   freeaddrinfo(res);
